refactor: Service단 변경으로 인한 수정
